+++ webpack.config.dev.js
@@ -26,7 +26,13 @@ module.exports = {
                 ],
                 options: {
                     presets: [
-                        '@babel/preset-env',
+                        ['@babel/preset-env', {
+                            useBuiltIns: 'usage',
+                            corejs: {
+                                version: '3.40',
+                            },
+                            targets: "defaults",
+                        }],
                     ],
                 },
             }