+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.8.0] - 2025-05-06
 ### Added
 - Added support for closing server-side banners with the state stored in a cookie.
 - Added support for animated closing. The first animation type `slideUp` implemented.
@@ -148,7 +150,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Added missing getter `Banner.resources`.
 
-[Unreleased]: https://github.com/68publishers/amp-client-js/compare/v1.7.0...HEAD
+[Unreleased]: https://github.com/68publishers/amp-client-js/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/68publishers/amp-client-js/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/68publishers/amp-client-js/compare/v1.6.1...v1.7.0
 [1.6.1]: https://github.com/68publishers/amp-client-js/compare/v1.6.0...v1.6.1
 [1.6.0]: https://github.com/68publishers/amp-client-js/compare/v1.5.0...v1.6.0